add new game logic
